Indicate test results via a button in reading mode
This adds a ToggleButton to the add-on.
In reading mode, an XHR request is made to the raw
version of a page. The button then shows the total
number of errors and warnings from testing against
the raw version. Clicking the button shows the
detailed results in the sidebar as usual.
When editing, the button is disabled, but will
show the result number as visual help.